+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 0.2.6 (unreleased)
+# 0.2.6
 - [#13] Meaningful exception types
 - [#20] Speed up Go Quartz process boot time
 - [#22] Handle large buffers